requirements.txt
numpy==1.19.2
Pillow==8.0.0
pywin32==228
requests==2.24.0